+++ apps/deploy-web/src/components/wallet/WalletModal/index.ts
@@ -1,1 +1,1 @@
-export { DefaultWalletModal, WalletModal } from "./WalletModal";
+export { WalletModal } from "./WalletModal";
