+++ apps/deploy-web/src/store/chainStore/chainStore.ts
@@ -1,23 +1,11 @@
-import { useEffect, useMemo } from "react";
-import type {
-  ChainContext,
-  ChainName,
-  DisconnectOptions,
-  MainWalletBase,
-  ManagerContext,
-  State,
-  WalletManager,
-  WalletRepo,
-  WalletStatus
-} from "@cosmos-kit/core";
+import type { MainWalletBase, WalletManager, WalletRepo } from "@cosmos-kit/core";
 import type { getDefaultStore } from "jotai";
-import { atom, useAtomValue } from "jotai";
+import { atom } from "jotai";
 
-import { getChainWalletContext } from "./utils";
 import { CURRENT_WALLET_KEY } from "./walletManagerFactory";
 import { loadAllWallets, loadWalletByName } from "./walletRegistry";
 
-interface ChainStoreOptions {
+export interface ChainStoreOptions {
   store: ReturnType<typeof getDefaultStore>;
 }
 
@@ -29,12 +17,12 @@ export class ChainStore {
   private readonly jotaiStore: ReturnType<typeof getDefaultStore>;
 
   readonly walletManagerAtom = atom<WalletManager | null>(null);
-  readonly stateVersionAtom = atom<number>(0);
-  readonly isInitializedAtom = atom<boolean>(false);
-  readonly isLoadingWalletsAtom = atom<boolean>(false);
-  readonly modalIsOpenAtom = atom<boolean>(false);
+  readonly stateVersionAtom = atom(0);
+  readonly isInitializedAtom = atom(false);
+  readonly isLoadingWalletsAtom = atom(false);
+  readonly modalIsOpenAtom = atom(false);
   readonly modalWalletRepoAtom = atom<WalletRepo | undefined>(undefined);
-  readonly selectedWalletNameAtom = atom<string>("");
+  readonly selectedWalletNameAtom = atom("");
 
   private allWalletsLoaded = false;
 
@@ -56,7 +44,7 @@ export class ChainStore {
       const wallets = await loadWalletByName(storedWallet);
       await this.replaceManager(wallets);
     } catch {
-      window.localStorage.removeItem(CURRENT_WALLET_KEY);
+      window.localStorage?.removeItem(CURRENT_WALLET_KEY);
     }
   }
 
@@ -65,7 +53,6 @@ export class ChainStore {
 
     try {
       const allWallets = await loadAllWallets();
-      console.log(allWallets);
       await this.replaceManager(allWallets);
       this.allWalletsLoaded = true;
       this.bumpVersion();
@@ -161,255 +148,4 @@ export class ChainStore {
   setSelectedWalletName(name: string): void {
     this.jotaiStore.set(this.selectedWalletNameAtom, name);
   }
-
-  getDisconnectedStub(chainName: ChainName): ChainContext {
-    return {
-      chainWallet: undefined,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      chain: undefined as any,
-      assets: undefined,
-      wallet: undefined,
-      logoUrl: undefined,
-      address: undefined,
-      username: undefined,
-      message: undefined,
-      status: "Disconnected" as WalletStatus.Disconnected,
-      isWalletDisconnected: true,
-      isWalletConnecting: false,
-      isWalletConnected: false,
-      isWalletRejected: false,
-      isWalletNotExist: false,
-      isWalletError: false,
-      connect: async () => {
-        await this.ensureAllWalletsLoaded();
-        const mgr = this.jotaiStore.get(this.walletManagerAtom);
-        if (mgr) {
-          const repo = mgr.getWalletRepo(chainName);
-          repo.openView();
-        }
-      },
-      disconnect: async () => {},
-      getRpcEndpoint: () => Promise.reject(new Error("Not connected")),
-      getRestEndpoint: () => Promise.reject(new Error("Not connected")),
-      getStargateClient: () => Promise.reject(new Error("Not connected")),
-      getCosmWasmClient: () => Promise.reject(new Error("Not connected")),
-      getSigningStargateClient: () => Promise.reject(new Error("Not connected")),
-      getSigningCosmWasmClient: () => Promise.reject(new Error("Not connected")),
-      getNameService: () => Promise.reject(new Error("Not connected")),
-      estimateFee: () => Promise.reject(new Error("Not connected")),
-      sign: () => Promise.reject(new Error("Not connected")),
-      broadcast: () => Promise.reject(new Error("Not connected")),
-      signAndBroadcast: () => Promise.reject(new Error("Not connected")),
-      qrUrl: undefined,
-      appUrl: undefined,
-      defaultSignOptions: undefined,
-      setDefaultSignOptions: () => {},
-      enable: () => Promise.reject(new Error("Not connected")),
-      suggestToken: () => Promise.reject(new Error("Not connected")),
-      getAccount: () => Promise.reject(new Error("Not connected")),
-      getOfflineSigner: () => {
-        throw new Error("Not connected");
-      },
-      getOfflineSignerAmino: () => {
-        throw new Error("Not connected");
-      },
-      getOfflineSignerDirect: () => {
-        throw new Error("Not connected");
-      },
-      signAmino: () => Promise.reject(new Error("Not connected")),
-      signDirect: () => Promise.reject(new Error("Not connected")),
-      signArbitrary: () => Promise.reject(new Error("Not connected")),
-      sendTx: () => Promise.reject(new Error("Not connected")),
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      walletRepo: undefined as any,
-      openView: () => {
-        this.ensureAllWalletsLoaded();
-      },
-      closeView: () => {}
-    };
-  }
-}
-
-// --- Hooks (standalone functions that close over a ChainStore instance) ---
-
-export function createChainStoreHooks(store: ChainStore) {
-  function useChain(chainName: ChainName): ChainContext {
-    const manager = useAtomValue(store.walletManagerAtom);
-    const version = useAtomValue(store.stateVersionAtom);
-
-    const context = useMemo(() => {
-      if (!manager) {
-        return store.getDisconnectedStub(chainName);
-      }
-
-      const walletRepo = manager.getWalletRepo(chainName);
-      walletRepo.activate();
-
-      const { disconnect, openView, closeView, current, chainRecord, getRpcEndpoint, getRestEndpoint, getStargateClient, getCosmWasmClient, getNameService } =
-        walletRepo;
-      const { chain, assetList } = chainRecord;
-
-      const chainWalletContext = chain ? getChainWalletContext(chain.chain_id, current, true) : undefined;
-
-      return {
-        ...chainWalletContext,
-        walletRepo,
-        chain,
-        assets: assetList,
-        openView,
-        closeView,
-        connect: async () => {
-          await store.ensureAllWalletsLoaded();
-          openView();
-        },
-        disconnect: (options?: DisconnectOptions) => disconnect(void 0, true, options),
-        getRpcEndpoint,
-        getRestEndpoint,
-        getStargateClient,
-        getCosmWasmClient,
-        getNameService
-      } as ChainContext;
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [manager, version, chainName]);
-
-    useEffect(() => {
-      if (!manager || !context.isWalletDisconnected) return;
-
-      const storedWallet = window.localStorage.getItem(CURRENT_WALLET_KEY);
-      if (storedWallet) {
-        const walletRepo = manager.getWalletRepo(chainName);
-        walletRepo.connect(storedWallet);
-      }
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [manager, chainName]);
-
-    return context;
-  }
-
-  function useManager(): ManagerContext {
-    const manager = useAtomValue(store.walletManagerAtom);
-    const version = useAtomValue(store.stateVersionAtom);
-
-    return useMemo(() => {
-      if (!manager) {
-        return {
-          chainRecords: [],
-          walletRepos: [],
-          mainWallets: [],
-          defaultNameService: "icns",
-          getChainRecord: () => {
-            throw new Error("WalletManager not initialized");
-          },
-          getWalletRepo: () => {
-            throw new Error("WalletManager not initialized");
-          },
-          addChains: () => {
-            throw new Error("WalletManager not initialized");
-          },
-          addEndpoints: () => {},
-          getChainLogo: () => undefined,
-          getNameService: () => {
-            throw new Error("WalletManager not initialized");
-          },
-          on: () => {},
-          off: () => {}
-        } as unknown as ManagerContext;
-      }
-
-      return {
-        chainRecords: manager.chainRecords,
-        walletRepos: manager.walletRepos,
-        mainWallets: manager.mainWallets,
-        defaultNameService: manager.defaultNameService,
-        getChainRecord: manager.getChainRecord,
-        getWalletRepo: manager.getWalletRepo,
-        addChains: manager.addChains,
-        addEndpoints: manager.addEndpoints,
-        getChainLogo: manager.getChainLogo,
-        getNameService: manager.getNameService,
-        on: manager.on,
-        off: manager.off
-      };
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [manager, version]);
-  }
-
-  function useWallet(walletName?: string) {
-    const manager = useAtomValue(store.walletManagerAtom);
-    const version = useAtomValue(store.stateVersionAtom);
-
-    return useMemo(() => {
-      if (!manager) {
-        return {
-          mainWallet: undefined,
-          chainWallets: [],
-          wallet: undefined,
-          status: "Disconnected" as WalletStatus.Disconnected,
-          message: undefined
-        };
-      }
-
-      const mainWallet: MainWalletBase | undefined = walletName
-        ? manager.getMainWallet(walletName)
-        : manager.mainWallets.find(w => w.isActive && w.clientMutable.state !== ("Error" as State));
-
-      if (!mainWallet) {
-        return {
-          mainWallet: undefined,
-          chainWallets: [],
-          wallet: undefined,
-          status: "Disconnected" as WalletStatus.Disconnected,
-          message: undefined
-        };
-      }
-
-      const { walletInfo, getChainWalletList, getGlobalStatusAndMessage } = mainWallet;
-      const [globalStatus, globalMessage] = getGlobalStatusAndMessage(true);
-
-      return {
-        mainWallet,
-        chainWallets: getChainWalletList(false),
-        wallet: walletInfo,
-        status: globalStatus,
-        message: globalMessage
-      };
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [manager, version, walletName]);
-  }
-
-  function useWalletClient(walletName?: string) {
-    const manager = useAtomValue(store.walletManagerAtom);
-    const version = useAtomValue(store.stateVersionAtom);
-
-    return useMemo(() => {
-      if (!manager) {
-        return {
-          client: undefined,
-          status: "Init" as State,
-          message: undefined
-        };
-      }
-
-      const mainWallet: MainWalletBase | undefined = walletName ? manager.getMainWallet(walletName) : manager.mainWallets.find(w => w.isActive);
-
-      if (!mainWallet) {
-        return {
-          client: undefined,
-          status: "Init" as State,
-          message: undefined
-        };
-      }
-
-      const { clientMutable } = mainWallet;
-
-      return {
-        client: clientMutable.data,
-        status: clientMutable.state,
-        message: clientMutable.message
-      };
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [manager, version, walletName]);
-  }
-
-  return { useChain, useManager, useWallet, useWalletClient };
 }
