+++ apps/deploy-web/src/store/chainStore/index.ts
@@ -1,16 +1,11 @@
 import { store } from "@src/store/global-store";
 import networkStore from "@src/store/networkStore";
-import { ChainStore, createChainStoreHooks } from "./chainStore";
+import { ChainStore } from "./chainStore";
+import { createChainStoreHooks } from "./hooks";
 
 export type { ChainContext } from "@cosmos-kit/core";
 
-const chainStore = ChainStore.create({ store });
-const { useChain, useManager, useWallet, useWalletClient } = createChainStoreHooks(chainStore);
+export const chainStore = ChainStore.create({ store });
+const { useChain, useManager, useWallet, useWalletClient, useSelectedChain } = createChainStoreHooks(chainStore, networkStore);
 
-export function useSelectedChain() {
-  const { chainRegistryName } = networkStore.useSelectedNetwork();
-  return useChain(chainRegistryName);
-}
-
-export { useChain, useManager, useWallet, useWalletClient };
-export default chainStore;
+export { useChain, useManager, useWallet, useWalletClient, useSelectedChain };
